Update mozphab version,json
  - usually this container will not be deployed as a standalone instance thus
the version.json is not needed for Dockerflow purposes. However, the version
information *is* important for builds that use this container as a
source. Make the version information available in the mozphab.json file
instead of version.json so later containers can create+update
version.json in their build.

we export /app at the phabext build, so don't export it here