Tweaks needed for GCP
Brief explanation: GCP maps config files into the containers in a
read-only mode via configMaps. For whatever reason, phabricator really
hates the idea of local.json being read-only and will complain loudly.

To reduce background noise, I'd like to map the local.json into a
holding directory and then copy it into the correct location prior to
application startup. This way, the file should A) Exist where it needs
to and B) Be writable by the phabricator user.

These changes should allow this with a minimum of fuss. Feel free to
follow up with any questions.